+++ include/section_list_loader.h
@@ -64,6 +64,13 @@ extern int locate_article_in_section(SECTION_LIST *p_section, const ARTICLE *p_a
 //           failure : error number (< 0)
 extern int scan_unread_article_in_section(SECTION_LIST *p_section, const ARTICLE *p_article_cur, const ARTICLE **pp_article_unread);
 
+extern int scan_article_in_section_by_uid(SECTION_LIST *p_section, const ARTICLE *p_article_cur,
+										  int direction, int32_t uid, const ARTICLE **pp_article);
+extern int scan_article_in_section_by_username(SECTION_LIST *p_section, const ARTICLE *p_article_cur,
+										  int direction, const char *username, const ARTICLE **pp_article);
+extern int scan_article_in_section_by_title(SECTION_LIST *p_section, const ARTICLE *p_article_cur,
+										  int direction, const char *title, const ARTICLE **pp_article);
+
 extern int get_section_ex_menu_set(SECTION_LIST *p_section, MENU_SET *p_ex_menu_set);
 
 #endif //_SECTION_LIST_LOADER_H_